Refactor, add service package.
