+++ src/main/java/jp/aha/oretama/typoChecker/configuration/TypoCheckerConfiguration.java
@@ -1,6 +1,6 @@
 package jp.aha.oretama.typoChecker.configuration;
 
-import jp.aha.oretama.typoChecker.BaseDictionaryService;
+import jp.aha.oretama.typoChecker.service.BaseDictionaryService;
 import jp.aha.oretama.typoChecker.language.CodingEnglish;
 import org.languagetool.JLanguageTool;
 import org.languagetool.rules.Rule;
